pyproject.toml
[project]
name = "balsam-lm-evaluation"
version = "0.1.0"
requires-python = ">=3.10,<4"
dependencies = [
    "lmms-eval[all]~=0.7.1",
    "pyyaml~=6.0.3",
    "google-cloud-storage>=2.0.0,<3.9.0",
    "six~=1.17.0",
    "nltk~=3.9.2",
    "pyarabic~=0.6.15",
    "deepeval~=3.8.6",
    "python-dotenv~=1.2.1",
    "google-genai~=1.61.0",
    "groq~=1.0.0",
    "absl-py~=2.4.0",
    "rouge-score~=0.1.2",
    "pillow~=12.2.0",
]

[dependency-groups]
dev = [
    "mypy~=1.20.0",
    "pylint[spelling]~=4.0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = [".", "src/"]

[tool.coverage.run]
omit = ["__init__.py", "tests/*", "*/tests/*"]

[tool.coverage.report]
skip_empty = true

[tool.yapf]
based_on_style = "google"
indent_width = 4

[tool.yapfignore]
ignore_patterns = ["venv", "build", "dist"]
